+++ Resource/Resource/Font.h
@@ -38,7 +38,7 @@ public:
 	
 	void setColor(float r, float g, float b, float a);
 	void setString(const TCHAR *string);
-	void setRect(float left, float top, float right, float bottom); //추후에 RectTransform 구현 후 사라질 예정
+	void setRect(long left, long top, long right, long bottom); //추후에 RectTransform 구현 후 사라질 예정
 	void draw(); // 나중에 component에서 상속 에정
 	void update(); // 나중에 component에서 상속 에정
 	void release(); // 나중에 component에서 상속 에정
